requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.20
pymysql==1.0.3
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv==1.0.0
PyJWT==2.10.0
